AI Updates 2
Found AI node movement bug and fixed it.

Can find the closest item spawn and moves to it. It can move to nodes
after getting an item, but it will only check for the Y position of
this. I'll need to find a way to fix this.

Added some stuff for finding enemies.